requirements.txt
addict==2.4.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiosignal==1.3.1
altair==5.5.0
async-timeout==5.0.1
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
docker-pycreds==0.4.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
humanfriendly==10.0
idna==3.10
igraph==0.11.8
imageio==2.36.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.7
lazy_loader==0.4
lightning==2.4.0
lightning-utilities==0.11.9
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.1.0
narwhals==1.42.0
networkx==3.4.2
numpy==2.1.3
onnx==1.17.0
onnxruntime==1.20.1
opencv-python==4.10.0.84
packaging==24.2
pandas==2.3.0
pillow==11.0.0
pip==22.2.1
platformdirs==4.3.6
propcache==0.2.0
protobuf==5.29.0
psutil==6.1.0
pyarrow==20.0.0
pycocotools==2.0.8
pycparser==2.22
pydeck==0.9.1
pyparsing==3.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
Rtree==1.3.0
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
segment-anything==1.0
sentry-sdk==2.19.0
setproctitle==1.3.4
setuptools==63.2.0
shapely==2.0.6
six==1.16.0
smmap==5.0.1
streamlit==1.45.1
sympy==1.13.1
tcod==16.2.3
tenacity==9.1.2
texttable==1.7.0
threadpoolctl==3.5.0
tifffile==2024.9.20
toml==0.10.2
torch==2.5.1+cu124
torchaudio==2.5.1+cu124
torchmetrics==1.6.0
torchvision==0.20.1
tornado==6.5.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.3
wandb==0.18.7
watchdog==6.0.0
yarl==1.18.0